--- ts/packages/agents/chat/src/config.json
@@ -1,7 +1,0 @@
-{
-  "lookup": {
-    "fastMode": true,
-    "fastModelName": "GPT_35_TURBO",
-    "documentConcurrency": 1
-  }
-}
