+++ ts/packages/defaultAgentProvider/test/data/translate-e2e.json
@@ -5,19 +5,19 @@
   },
   {
     "request": "I am planning a trip to Spain. can you look up some of the cities I should visit.",
-    "action": "chat.lookupAndAnswer"
+    "action": "dispatcher.lookup.lookupAndAnswer"
   },
   {
     "request": "Look up the ingredients for lemon meringue pie",
-    "action": "chat.lookupAndAnswer"
+    "action": "dispatcher.lookup.lookupAndAnswer"
   },
   {
     "request": "Did duke win its bball game",
-    "action": "chat.lookupAndAnswer"
+    "action": "dispatcher.lookup.lookupAndAnswer"
   },
   {
     "request": "Lookup duke result on the web",
-    "action": "chat.lookupAndAnswer"
+    "action": "dispatcher.lookup.lookupAndAnswer"
   },
   { "request": "add play to the To do list", "action": "list.addItems" },
   { "request": "add homework to To do list", "action": "list.addItems" },
@@ -60,6 +60,20 @@
       "player.getFavorites",
       "player.filterTracks",
       "player.createPlaylist"
+    ]
+  },
+  {
+    "request": "look up information about Oregon and write a poem using what you found",
+    "action": [
+      "dispatcher.lookup.lookupAndAnswer",
+      "dispatcher.pendingRequestAction"
+    ]
+  },
+  {
+    "request": "look up the recipe for tarte a l'oignan and add the ingredients to my grocery list",
+    "action": [
+      "dispatcher.lookup.lookupAndAnswer",
+      "dispatcher.pendingRequestAction"
     ]
   }
 ]
