+++ ts/packages/defaultAgentProvider/test/data/translate-lookup-e2e.json
@@ -1,0 +1,24 @@
+[
+  {
+    "request": "add the ingredients for chess pie to the grocery list",
+    "action": [
+      "dispatcher.lookup.lookupAndAnswer",
+      "dispatcher.pendingRequestAction"
+    ]
+  },
+  {
+    "request": "play this week's top 10 on the billboard chart",
+    "action": [
+      "dispatcher.lookup.lookupAndAnswer",
+      "dispatcher.pendingRequestAction"
+    ]
+  },
+  {
+    "request": "play the song by Bach we talked about yesterday",
+    "action": "dispatcher.clarify.clarifyUnresolvedReference"
+  },
+  {
+    "request": "play song from the artist that we talked about when we were discussing flower",
+    "action": "dispatcher.clarify.clarifyUnresolvedReference"
+  }
+]
