+++ ts/packages/defaultAgentProvider/test/translate_lookup.test.ts
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+/*
+ * NOTE: This test has the extension .test.ts and not .spec.ts.
+ * *.test.ts files are run under test:live && test:live:debug
+ * project settings (see ../package.json).  The assumption is
+ * test:live has API endpoints where as test:local tests run
+ * wholly locally.
+ */
+
+import { defineTranslateTest } from "./translateTestCommon.js";
+const dataFiles = ["test/data/translate-lookup-e2e.json"];
+
+await defineTranslateTest("translate (w/implicit lookup)", dataFiles);
