+++ ts/packages/dispatcher/data/internetSearchConfig.json
@@ -1,0 +1,8 @@
+{
+    "lookup": {
+      "fastMode": true,
+      "fastModelName": "GPT_35_TURBO",
+      "documentConcurrency": 1
+    }
+  }
+
