+++ ts/packages/dispatcher/src/context/dispatcher/dispatcherAgent.ts
@@ -17,7 +17,13 @@ import {
 import { CommandHandlerContext } from "../commandHandlerContext.js";
 import { createActionResultNoDisplay } from "@typeagent/agent-sdk/helpers/action";
 import { DispatcherActions } from "./schema/dispatcherActionSchema.js";
-import { ClarifyRequestAction } from "./schema/clarifyActionSchema.js";
+import {
+    ClarifyRequestAction,
+    ClarifyUnresolvedReference,
+} from "./schema/clarifyActionSchema.js";
+import { loadAgentJsonTranslator } from "../../translation/agentTranslators.js";
+import { lookupAndAnswer } from "../../search/search.js";
+import { LookupAndAnswerAction } from "./schema/lookupActionSchema.js";
 
 const dispatcherHandlers: CommandHandlerTable = {
     description: "Type Agent Dispatcher Commands",
@@ -29,15 +35,27 @@ const dispatcherHandlers: CommandHandlerTable = {
 };
 
 async function executeDispatcherAction(
-    action: TypeAgentAction<DispatcherActions | ClarifyRequestAction>,
+    action: TypeAgentAction<
+        DispatcherActions | ClarifyRequestAction | LookupAndAnswerAction
+    >,
     context: ActionContext<CommandHandlerContext>,
 ) {
-    if (
-        action.actionName === "clarifyMultiplePossibleActionName" ||
-        action.actionName === "clarifyMissingParameter" ||
-        action.actionName === "clarifyUnresolvedReference"
-    ) {
-        return clarifyRequestAction(action, context);
+    switch (action.translatorName) {
+        case "dispatcher.clarify":
+            switch (action.actionName) {
+                case "clarifyMultiplePossibleActionName":
+                case "clarifyMissingParameter":
+                    return clarifyRequestAction(action, context);
+                case "clarifyUnresolvedReference":
+                    return clarifyUnresolvedReferenceAction(action, context);
+            }
+            break;
+        case "dispatcher.lookup":
+            switch (action.actionName) {
+                case "lookupAndAnswer":
+                    return lookupAndAnswer(action, context);
+            }
+            break;
     }
 
     throw new Error(`Unknown dispatcher action: ${action.actionName}`);
@@ -61,6 +79,46 @@ function clarifyRequestAction(
     return result;
 }
 
+async function clarifyUnresolvedReferenceAction(
+    action: ClarifyUnresolvedReference,
+    context: ActionContext<CommandHandlerContext>,
+) {
+    const agents = context.sessionContext.agentContext.agents;
+    if (
+        !agents.isSchemaActive("dispatcher.lookup") ||
+        !agents.isActionActive("dispatcher.lookup")
+    ) {
+        // lookup is disabled either for translation or action. Just ask the user.
+        return clarifyRequestAction(action, context);
+    }
+
+    const actionConfigs = [
+        agents.getActionConfig("dispatcher.lookup"),
+        agents.getActionConfig("dispatcher.clarify"),
+        agents.getActionConfig("dispatcher"),
+    ];
+    // TODO: cache this?
+    const translator = loadAgentJsonTranslator(
+        actionConfigs,
+        [],
+        agents,
+        false, // no multiple
+    );
+
+    const result = await translator.translate(
+        `What is ${action.parameters.reference}?`,
+    );
+
+    if (result.success) {
+        const action = result.data;
+        if (action.actionName === "lookupAndAnswer") {
+            return lookupAndAnswer(action as LookupAndAnswerAction, context);
+        }
+    }
+
+    return clarifyRequestAction(action, context);
+}
+
 export const dispatcherManifest: AppAgentManifest = {
     emojiChar: "🤖",
     description: "Built-in agent to dispatch requests",
@@ -78,6 +136,17 @@ export const dispatcherManifest: AppAgentManifest = {
                 schemaFile:
                     "./src/context/dispatcher/schema/clarifyActionSchema.ts",
                 schemaType: "ClarifyRequestAction",
+                injected: true,
+                cached: false,
+            },
+        },
+        lookup: {
+            schema: {
+                description:
+                    "Action that helps you look up information to answer user questions.",
+                schemaFile:
+                    "./src/context/dispatcher/schema/lookupActionSchema.ts",
+                schemaType: "LookupAction",
                 injected: true,
                 cached: false,
             },
